requirements.txt
requests
beautifulsoup4
faker
lxml
apscheduler
python-dotenv